Check permissions in object context when issuing commands

+++ application/forms/Command/Object/DeleteCommentForm.php
@@ -38,6 +38,13 @@ class DeleteCommentForm extends CommandForm
 
     protected function getCommand(Model $object)
     {
+        if (! $this->isGrantedOn(
+            'monitoring/command/comment/delete',
+            $object->{$object->object_type}
+        )) {
+            return null;
+        }
+
         $command = new DeleteCommentCommand();
         $command->setCommentName($object->name);
         $command->setAuthor($this->getAuth()->getUser()->getUsername());
